+++ cliboa/adapter/file.py
